Add example to extract the debug info from llvm ir

+++ src/expr_llvm/test.c
@@ -1,7 +1,7 @@
 
 
 typedef struct {
-    int a;
+    double a;
     int b;
 } my_struct;
